[mod] Fixture.java int to Long

+++ src/main/java/com/ek/earlykross/entity/Fixture.java
@@ -26,7 +26,7 @@ import lombok.ToString;
 public class Fixture {
 
     @Id // PK 지정
-    private int fId; // 경기번호
+    private Long fId; // 경기번호
     
     @ManyToOne
     @JoinColumn(name = "home", nullable = false)
